// landing/src/app/modules/setting/models/dashboard-setting.ts
import { BaseEntityTrace } from 'src/app/_metronic/models/base-entity-trace';
export interface DashboardSetting extends BaseEntityTrace {
  displayName: string;
  logoPath: string;
  smallLogoPath: string;
  bsPrimary: string;
  bsSecondPrimary: string;
  landingPrimary: string;
  landingSecondPrimary: string;
  bsPrimaryInverse: string;
  phones: string;
  address: string;
  homeImagePath: string;
  aboutImagePath: string;
  serviceImagePath: string;
  isOpenLandindOne: boolean;
  isOpenLandindTwo: boolean;
  landingMainPath: string;
  landingAboutPath: string;
  landingJoinUsPath: string;
  isLogoImage: boolean;
  pageTitle :string;
  pageDescription :string;
  robots :string;
  canonical :string;
  ogLocale :string;
  ogSiteName :string;
  ogType :string;
  ogTitle :string;
  ogDescription :string;
  ogUrl :string;
  twitterCard :string;
  twitterTitle :string;
  twitterDescription :string;
  email:string;
  facebookUrl:string;
  youtubeUrl:string;
  instgramUrl:string;
  linkedinUrl:string;
  whatsUrl:string;
}